Renaming utils.py
